+++ CHANGELOG.md
@@ -73,6 +73,7 @@ This major relase contains various fixes, changes and new features including:
 * CORE: [FEATURE] It is now possible to set a specific share name #94 
 * CORE: [FIX] Its not possible anymore to set invalid meta data #108
 * CORE: [CHANGE] Shares are now updated during every identy set (after authentication) instead on specific requests
+* CORE: [CHANGE] Meta hex colors are replaced with a generic color name #124
 * API: [CHANGE] New API version v2 available, v1 is still fully functional, but new features are only available in v2. See upgrade guide.
 * API: [FEATURE] Implemented new endpoint GET /api/v2/desktop-client?format=format to fetch desktop client
 * API: [CHANGE] removed GET /api/v2/about